website_sys/userlooker_sys/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pymongo
ijson
boto3
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
httpx
slowapi
